backend/requirements_sadtalker.txt
# SadTalker Requirements
# Python dependencies for AI Avatar with Lip Sync

# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0
numpy>=1.23.0
opencv-python>=4.7.0
pillow>=9.4.0
scipy>=1.10.0
scikit-image>=0.19.0
imageio>=2.25.0
imageio-ffmpeg>=0.4.8
librosa>=0.10.0

# Face detection and analysis
face-alignment>=1.3.5
dlib>=19.24.0

# Audio processing
soundfile>=0.12.0
pydub>=0.25.1

# Flask server
flask>=2.3.0
flask-cors>=4.0.0

# Text-to-Speech
gTTS>=2.3.0
pyttsx3>=2.90

# SadTalker specific
basicsr>=1.4.2
facexlib>=0.3.0
realesrgan>=0.3.0
gfpgan>=1.3.8

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
yacs>=0.1.8
requests>=2.28.0

# Image enhancement (optional but recommended)
lpips>=0.1.4
kornia>=0.6.12
